+++ attmap/__init__.py
@@ -2,11 +2,13 @@
 
 from .attmap import AttMap
 from .attmap_echo import AttMapEcho
-from .pathex_attmap import PathExAttMap
+from .helpers import *
+from .ordattmap import OrdAttMap
+from .ordpathex_attmap import OrdPathExAttMap
 from ._version import __version__
 
 AttributeDict = AttMap
 AttributeDictEcho = AttMapEcho
 
 __all__ = ["AttMap", "AttMapEcho", "AttributeDict", "AttributeDictEcho",
-           "PathExAttMap"]
+           "OrdAttMap", "OrdPathExAttMap", "get_data_lines"]