+++ attmap/_att_map_like.py
@@ -6,7 +6,7 @@ if sys.version_info < (3, 3):
     from collections import Mapping, MutableMapping
 else:
     from collections.abc import Mapping, MutableMapping
-from .helpers import get_logger
+from .helpers import is_custom_map, get_data_lines, get_logger
 
 __author__ = "Vince Reuter"
 __email__ = "vreuter@virginia.edu"
@@ -29,69 +29,28 @@ class AttMapLike(MutableMapping):
         """
         self.add_entries(entries)
 
-    @abc.abstractmethod
     def __getattr__(self, item, default=None):
-        pass
+        try:
+            return super(AttMapLike, self).__getattribute__(item)
+        except AttributeError:
+            try:
+                return self.__getitem__(item)
+            except KeyError:
+                # Requested item is unknown, but request was made via
+                # __getitem__ syntax, not attribute-access syntax.
+                raise AttributeError(item)
 
     @abc.abstractmethod
     def __setitem__(self, key, value):
         pass
 
+    @abc.abstractmethod
     def __getitem__(self, item):
-        try:
-            # Ability to return requested item name itself is delegated.
-            return self.__getattr__(item)
-        except AttributeError:
-            # Requested item is unknown, but request was made via
-            # __getitem__ syntax, not attribute-access syntax.
-            raise KeyError(item)
+        pass
 
+    @abc.abstractmethod
     def __delitem__(self, item):
-        try:
-            del self.__dict__[item]
-        except KeyError:
-            _LOGGER.debug("No item {} to delete".format(item))
-
-    def __eq__(self, other):
-        # TODO: check for equality across classes?
-        if not isinstance(other, Mapping):
-            return False
-        if len(self) != len(other):
-            # Ensure we don't have to worry about other containing self.
-            return False
-        for k, v in self.items():
-            if self._omit_from_eq(k):
-                _LOGGER.debug("Excluding from comparison: {}".format(k))
-                continue
-            try:
-                if not self._cmp(v, other[k]):
-                    return False
-            except KeyError:
-                return False
-        return True
-
-    @staticmethod
-    def _cmp(a, b):
-        def same_type(obj1, obj2, typenames=None):
-            t1, t2 = str(obj1.__class__), str(obj2.__class__)
-            return (t1 in typenames and t2 in typenames) if typenames else t1 == t2
-        if same_type(a, b, ["<type 'numpy.ndarray'>",
-                            "<class 'numpy.ndarray'>"]) or \
-            same_type(a, b, ["<class 'pandas.core.series.Series'>"]):
-            check = lambda x, y: (x == y).all()
-        elif same_type(a, b, ["<class 'pandas.core.frame.DataFrame'>"]):
-            check = lambda x, y: (x == y).all().all()
-        else:
-            check = lambda x, y: x == y
-        try:
-            return check(a, b)
-        except ValueError:
-            # ValueError arises if, e.g., the pair of Series have
-            # have nonidentical labels.
-            return False
-
-    def __ne__(self, other):
-        return not self == other
+        pass
 
     def __iter__(self):
         return iter([k for k in self.__dict__.keys()])
@@ -100,11 +59,13 @@ class AttMapLike(MutableMapping):
         return sum(1 for _ in iter(self))
 
     def __repr__(self):
-        return repr({k: v for k, v in self.__dict__.items()
-                    if not self._omit_from_repr(k, self.__class__)})
-
-    def __str__(self):
-        return "{}: {}".format(self.__class__.__name__, repr(self))
+        base = self.__class__.__name__
+        data = self._simplify_keyvalue(self._data_for_repr())
+        if data:
+            return base + "\n" + "\n".join(
+                get_data_lines(data, lambda obj: repr(obj).strip("'")))
+        else:
+            return base + ": {}"
 
     def add_entries(self, entries):
         """
@@ -113,23 +74,21 @@ class AttMapLike(MutableMapping):
         :param Iterable[(object, object)] | Mapping | pandas.Series entries:
             collection of pairs of keys and values
         """
-        if not entries:
+        if entries is None:
             return
         # Permit mapping-likes and iterables/generators of pairs.
         if callable(entries):
             entries = entries()
-        elif "pandas.core.series.Series" in type(entries).__bases__:
+        elif any("pandas.core" in str(t) for t in type(entries).__bases__):
             entries = entries.to_dict()
         try:
             entries_iter = entries.items()
         except AttributeError:
             entries_iter = entries
         for k, v in entries_iter:
-            if k not in self or not \
-                    (isinstance(v, Mapping) and isinstance(self[k], AttMapLike)):
-                self[k] = v
-            else:
-                self[k] = self[k].add_entries(v)
+            self[k] = v if k not in self or not \
+                (isinstance(v, Mapping) and isinstance(self[k], AttMapLike)) \
+                else self[k].add_entries(v)
         return self
 
     def is_null(self, item):
@@ -154,20 +113,22 @@ class AttMapLike(MutableMapping):
         """
         Convert this instance to a dict.
 
-        :return dict[str, object]:
+        :return dict[str, object]: this map's data, in a simpler container
         """
-        def go(kvs, acc):
-            try:
-                h, t = kvs[0], kvs[1:]
-            except IndexError:
-                return acc
-            k, v = h
-            acc[k] = go(list(v.items()), {}) \
-                if isinstance(v, Mapping) and not isinstance(v, dict) else v
-            return go(t, acc)
-        return go(list(self.items()), {})
+        return self._simplify_keyvalue(
+            self.items(), self._new_empty_basic_map())
 
-    def _omit_from_eq(self, k):
+    def _data_for_repr(self):
+        """
+        Hook for extracting the data used in the object's text representation.
+
+        :return Iterable[(hashable, object)]: collection of key-value pairs
+            to include in object's text representation
+        """
+        return filter(lambda kv: not self._excl_from_repr(kv[0], self.__class__),
+                      self.items())
+
+    def _excl_from_eq(self, k):
         """
         Hook for exclusion of particular value from a representation
 
@@ -176,7 +137,7 @@ class AttMapLike(MutableMapping):
         """
         return False
 
-    def _omit_from_repr(self, k, cls):
+    def _excl_from_repr(self, k, cls):
         """
         Hook for exclusion of particular value from a representation
 
@@ -186,3 +147,31 @@ class AttMapLike(MutableMapping):
             text representation
         """
         return False
+
+    @abc.abstractproperty
+    def _lower_type_bound(self):
+        """ Most specific type to which stored Mapping should be transformed """
+        pass
+
+    @abc.abstractmethod
+    def _new_empty_basic_map(self):
+        """ Return the empty collection builder for Mapping type simplification. """
+        pass
+
+    def _simplify_keyvalue(self, kvs, acc=None):
+        """
+        Simplify a collection of key-value pairs, "reducing" to simpler types.
+
+        :param Iterable[(object, object)] kvs: collection of key-value pairs
+        :param Iterable acc: accumulating collection of simplified data
+        :return Iterable: collection of simplified data
+        """
+        acc = acc or self._new_empty_basic_map()
+        kvs = iter(kvs)
+        try:
+            k, v = next(kvs)
+        except StopIteration:
+            return acc
+        acc[k] = self._simplify_keyvalue(
+            v.items(), self._new_empty_basic_map()) if is_custom_map(v) else v
+        return self._simplify_keyvalue(kvs, acc)