+++ attmap/attmap.py
@@ -6,8 +6,11 @@ if sys.version_info < (3, 3):
 else:
     from collections.abc import Mapping
 
-from .helpers import copy
+from .helpers import copy, get_logger, safedel_message
 from ._att_map_like import AttMapLike
+
+
+_LOGGER = get_logger(__name__)
 
 
 @copy
@@ -19,52 +22,61 @@ class AttMap(AttMapLike):
     facilitating attribute traversal (e.g., attmap.attr.attr).
     """
 
-    def __getattr__(self, item, default=None):
-        """
-        Fetch the value associated with the provided identifier.
-
-        :param int | str item: identifier for value to fetch
-        :return object: whatever value corresponds to the requested key/item
-        :raises AttributeError: if the requested item has not been set,
-            no default value is provided, and this instance is not configured
-            to return the requested key/item itself when it's missing; also,
-            if the requested item is unmapped and appears to be protected,
-            i.e. by flanking double underscores, then raise AttributeError
-            anyway. More specifically, respect attribute naming that appears
-            to be indicative of the intent of protection.
-        """
+    def __delitem__(self, key):
         try:
-            return super(AttMap, self).__getattribute__(item)
-        except (AttributeError, TypeError):
-            # Handle potential failure from non-string or property request.
-            pass
-        try:
-            # Route this dot notation request through the Mapping route.
-            return self.__dict__[item]
+            del self.__dict__[key]
         except KeyError:
-            # If not, triage and cope accordingly.
-            if item.startswith("__") and item.endswith("__"):
-                # Accommodate security-through-obscurity approach used by some libraries.
-                error_reason = "Protected-looking attribute: {}".format(item)
-                raise AttributeError(error_reason)
-            if default is not None:
-                # For compatibility with ordinary getattr() call, allow default value.
-                return default
-            # Throw up our hands in despair and resort to exception behavior.
-            raise AttributeError(item)
+            _LOGGER.debug(safedel_message(key))
 
-    @property
-    def _transformations(self):
-        """
-        Add path expansion behavior to more general attmap.
+    def __getitem__(self, item):
+        return self.__dict__[item]
 
-        :return list[(function, function)]: pairs in which first component is a
-            predicate and second is a function to apply to a value if it
-            satisfies the predicate
+    def __setitem__(self, key, value):
         """
-        newmap = lambda obj: isinstance(obj, Mapping) and \
-                             not isinstance(obj, self._lower_type_bound)
-        return [(newmap, self._metamorph_maplike)]
+        This is the key to making this a unique data type.
+
+        :param str key: name of the key/attribute for which to establish value
+        :param object value: value to which set the given key; if the value is
+            a mapping-like object, other keys' values may be combined.
+        """
+        # TODO: consider enforcement of type constraint, that value of different
+        # type may not overwrite existing.
+        self.__dict__[key] = self._finalize_value(value)
+
+    def __eq__(self, other):
+        # TODO: check for equality across classes?
+        if (type(self) != type(other)) or (len(self) != len(other)):
+            return False
+        for k, v in self.items():
+            if self._excl_from_eq(k):
+                _LOGGER.debug("Excluding from comparison: {}".format(k))
+                continue
+            if not self._cmp(v, other[k]):
+                return False
+        return True
+
+    def __ne__(self, other):
+        return not self == other
+
+    @staticmethod
+    def _cmp(a, b):
+        """ Hook to tailor value comparison in determination of map equality. """
+        def same_type(obj1, obj2, typenames=None):
+            t1, t2 = str(obj1.__class__), str(obj2.__class__)
+            return (t1 in typenames and t2 in typenames) if typenames else t1 == t2
+        if same_type(a, b, ["<type 'numpy.ndarray'>", "<class 'numpy.ndarray'>"]) or \
+            same_type(a, b, ["<class 'pandas.core.series.Series'>"]):
+            check = lambda x, y: (x == y).all()
+        elif same_type(a, b, ["<class 'pandas.core.frame.DataFrame'>"]):
+            check = lambda x, y: (x == y).all().all()
+        else:
+            check = lambda x, y: x == y
+        try:
+            return check(a, b)
+        except ValueError:
+            # ValueError arises if, e.g., the pair of Series have
+            # have nonidentical labels.
+            return False
 
     def _finalize_value(self, v):
         """
@@ -77,6 +89,10 @@ class AttMap(AttMapLike):
             if p(v):
                 return f(v)
         return v
+
+    @property
+    def _lower_type_bound(self):
+        return AttMap
 
     def _metamorph_maplike(self, m):
         """
@@ -93,36 +109,19 @@ class AttMap(AttMapLike):
         m_prime.__init__(m)
         return m_prime
 
-    def __setitem__(self, key, value):
-        """
-        This is the key to making this a unique data type. Flag set at
-        time of construction determines whether it's possible for a null
-        value to squash a non-null value. The combination of that flag and
-        one indicating whether request for value for unset attribute should
-        return the attribute name itself determines if any attribute/key
-        may be set to a null value.
-
-        :param str key: name of the key/attribute for which to establish value
-        :param object value: value to which set the given key; if the value is
-            a mapping-like object, other keys' values may be combined.
-        """
-        # TODO: consider enforcement of type constraint, that value of different
-        # type may not overwrite existing.
-        self.__dict__[key] = self._finalize_value(value)
-
-    def __getitem__(self, item):
-        try:
-            # Ability to return requested item name itself is delegated.
-            return self.__getattr__(item)
-        except AttributeError:
-            # Requested item is unknown, but request was made via
-            # __getitem__ syntax, not attribute-access syntax.
-            raise KeyError(item)
-
-    def __str__(self):
-        return "{}: {}".format(self.__class__.__name__, repr(self))
+    def _new_empty_basic_map(self):
+        """ Return the empty collection builder for Mapping type simplification. """
+        return dict()
 
     @property
-    def _lower_type_bound(self):
-        """ Most specific type to which stored Mapping should be transformed """
-        return AttMap
+    def _transformations(self):
+        """
+        Add path expansion behavior to more general attmap.
+
+        :return list[(function, function)]: pairs in which first component is a
+            predicate and second is a function to apply to a value if it
+            satisfies the predicate
+        """
+        newmap = lambda obj: isinstance(obj, Mapping) and \
+                             not isinstance(obj, self._lower_type_bound)
+        return [(newmap, self._metamorph_maplike)]